% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_to_col}
\alias{add_to_col}
\title{Pass matrix by reference and edit the first column.}
\usage{
add_to_col(x, col, y)
}
\arguments{
\item{x}{Integer matrix}

\item{col}{Index of col to modify}

\item{y}{value to add}
}
\description{
Pass matrix by reference and edit the first column.
}
