% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_to_row}
\alias{add_to_row}
\title{Pass matrix by reference and edit the first row.}
\usage{
add_to_row(x, row, y)
}
\arguments{
\item{x}{Integer matrix}

\item{row}{Index of row to modify}

\item{y}{value to add}
}
\description{
Pass matrix by reference and edit the first row.
}
